first step towards nodeification
