+++ js/index.js
@@ -1,0 +1,6 @@
+global.environment = require('./interface/environment')
+
+var controller = require('./mvc/controller')
+
+controller.init()
+controller.bindEventListeners()
