+++ js/services/get-random.js
@@ -1,3 +1,5 @@
 function getRandom (min, max) {
   return Math.random() * (max - min) + min;
 }
+
+module.exports = getRandom
