+++ package.json
@@ -2,7 +2,7 @@
   "name": "dREAM-cACHER_interface_1.0",
   "version": "1.0.0",
   "description": "This is the first iteration of the interface for the 'dREAM-cACHER' web application",
-  "main": "./js/mvc/controller.js",
+  "main": "./js/index.js",
   "scripts": {
     "start": "budo . -d --serve bundle.js --live",
     "build": "browserify . -o bundle.js -g uglifyify",
